# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cases/{caseId}/review": {
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Required (docs/03 §7.2): a retry with the same key and body returns the first result instead of a second review.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success envelope (docs/01 §7).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request: VALIDATION_ERROR, IDEMPOTENCY_KEY_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: UNAUTHORIZED, TOKEN_EXPIRED, AUTH_SESSION_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found: NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: REVIEW_CASE_NOT_CLOSED, REVIEW_NO_ACCEPTED_BID, REVIEW_ALREADY_EXISTS, IDEMPOTENCY_KEY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "426": {
            "description": "Upgrade required: APP_UPDATE_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests: RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal error: INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Review the attorney of a closed case (client, once per case)",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{id}": {
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success envelope (docs/01 §7).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request: VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: UNAUTHORIZED, TOKEN_EXPIRED, AUTH_SESSION_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found: NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: REVIEW_EDIT_WINDOW_EXPIRED, REVIEW_NOT_EDITABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "426": {
            "description": "Upgrade required: APP_UPDATE_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests: RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal error: INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit own review within review.edit_window_days (client)",
        "tags": [
          "reviews"
        ]
      }
    },
    "/attorneys/{id}/reviews": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Attorney user id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "meta.nextCursor of the previous page.",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope (docs/01 §7).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicReviewListEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request: VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: UNAUTHORIZED, TOKEN_EXPIRED, AUTH_SESSION_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found: NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "426": {
            "description": "Upgrade required: APP_UPDATE_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests: RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal error: INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Published reviews of an attorney, newest first",
        "tags": [
          "reviews"
        ]
      }
    },
    "/attorneys/{id}/reviews/summary": {
      "get": {
        "operationId": "summary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Attorney user id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success envelope (docs/01 §7).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSummaryEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request: VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: UNAUTHORIZED, TOKEN_EXPIRED, AUTH_SESSION_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found: NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "426": {
            "description": "Upgrade required: APP_UPDATE_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests: RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal error: INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Average rating, count and per-star distribution of an attorney",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{id}/report": {
      "post": {
        "operationId": "report",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Resource-creating POST: a retry with the same key and body is applied once (see IdempotencyInterceptor).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success envelope (docs/01 §7).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewReportEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request: VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: UNAUTHORIZED, TOKEN_EXPIRED, AUTH_SESSION_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found: NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: IDEMPOTENCY_KEY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "426": {
            "description": "Upgrade required: APP_UPDATE_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests: RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal error: INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Report a review to moderation (the reviewed attorney)",
        "tags": [
          "reviews"
        ]
      }
    }
  },
  "info": {
    "title": "LawBid API",
    "description": "LawBid — US legal-services marketplace. Every 2xx JSON body is the envelope {data, meta?: {nextCursor}} and every error is {error: {code, message, details?, requestId}} with `code` from the ErrorCode enum — docs/01_FOUNDATION_AUTH.md §7.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorCode": {
        "type": "string",
        "enum": [
          "VALIDATION_ERROR",
          "INTERNAL_ERROR",
          "NOT_FOUND",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RATE_LIMITED",
          "IDEMPOTENCY_KEY_REQUIRED",
          "IDEMPOTENCY_KEY_CONFLICT",
          "APP_UPDATE_REQUIRED",
          "TOKEN_EXPIRED",
          "AUTH_OTP_INVALID",
          "AUTH_OTP_EXPIRED",
          "AUTH_OTP_LOCKED",
          "AUTH_OTP_REQUEST_LIMIT",
          "AUTH_REFRESH_INVALID",
          "AUTH_REFRESH_EXPIRED",
          "AUTH_REFRESH_REUSE_DETECTED",
          "AUTH_SESSION_REVOKED",
          "AUTH_SOCIAL_TOKEN_INVALID",
          "AUTH_SOCIAL_PROVIDER_UNAVAILABLE",
          "AUTH_PROVIDER_DISABLED",
          "ACCOUNT_EXISTS_USE_OTHER_METHOD",
          "IDENTIFIER_ALREADY_LINKED",
          "CONTACT_DOMAIN_BLOCKED",
          "CONTACT_ALREADY_EXISTS",
          "REAUTH_REQUIRED",
          "REAUTH_INVALID",
          "DEVICE_ATTESTATION_REQUIRED",
          "ACCOUNT_SUSPENDED",
          "ACCOUNT_DELETED",
          "CLIENT_CONTACTS_INCOMPLETE",
          "ONBOARDING_INCOMPLETE",
          "ROLE_ALREADY_SET",
          "NOT_IMPLEMENTED",
          "I18N_LANGUAGE_NOT_FOUND",
          "I18N_IMPORT_INVALID",
          "PROVIDER_BUDGET_EXCEEDED",
          "PHONE_COUNTRY_NOT_SUPPORTED",
          "REVIEW_CASE_NOT_CLOSED",
          "REVIEW_NO_ACCEPTED_BID",
          "REVIEW_ALREADY_EXISTS",
          "REVIEW_EDIT_WINDOW_EXPIRED",
          "REVIEW_NOT_EDITABLE"
        ]
      },
      "ErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "Human-readable, not localized and not stable — never branch on it."
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "description": "Code-specific context, e.g. `validation` (VALIDATION_ERROR) or `missing` (CLIENT_CONTACTS_INCOMPLETE)."
          },
          "requestId": {
            "type": "string",
            "description": "X-Request-Id of the failed request."
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "ResponseMetaDto": {
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Opaque cursor for the next page; null on the last page."
          }
        }
      },
...
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "published",
          "hidden",
          "removed"
        ]
      },
      "ReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "authorDisplayName": {
            "type": "string",
            "nullable": true,
            "description": "Reviewer as \"First L.\" (docs/03 §7.4); null when the account has no name (show a localized placeholder).",
            "example": "Anna K."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "editedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "Set when the client edited the review (\"Edited\" label)."
          },
          "caseId": {
            "type": "string",
            "format": "uuid"
          },
          "attorneyId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewStatus"
              }
            ]
          },
          "editableUntil": {
            "type": "string",
            "format": "date-time",
            "description": "Last moment the client may edit (created + review.edit_window_days)."
          }
        },
        "required": [
          "id",
          "rating",
          "body",
          "authorDisplayName",
          "createdAt",
          "editedAt",
          "caseId",
          "attorneyId",
          "status",
          "editableUntil"
        ]
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "body": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000
          }
        },
        "required": [
          "rating"
        ]
      },
      "ReviewEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReviewDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateReviewDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "body": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000
          }
        }
      },
      "PublicReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "authorDisplayName": {
            "type": "string",
            "nullable": true,
            "description": "Reviewer as \"First L.\" (docs/03 §7.4); null when the account has no name (show a localized placeholder).",
            "example": "Anna K."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "editedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "Set when the client edited the review (\"Edited\" label)."
          }
        },
        "required": [
          "id",
          "rating",
          "body",
          "authorDisplayName",
          "createdAt",
          "editedAt"
        ]
      },
      "PublicReviewListEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicReviewDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "RatingBucketDto": {
        "type": "object",
        "properties": {
          "stars": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "stars",
          "count"
        ]
      },
      "ReviewSummaryDto": {
        "type": "object",
        "properties": {
          "ratingAvg": {
            "type": "number",
            "nullable": true,
            "description": "Average of published reviews rounded to one decimal; null when there are none (\"New — no reviews\").",
            "example": 4.5
          },
          "ratingCount": {
            "type": "integer"
          },
          "distribution": {
            "description": "Count per star, always 5 entries, 5 → 1.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RatingBucketDto"
            }
          }
        },
        "required": [
          "ratingAvg",
          "ratingCount",
          "distribution"
        ]
      },
      "ReviewSummaryEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReviewSummaryDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "ReportReason": {
        "type": "string",
        "enum": [
          "spam",
          "abuse",
          "misinformation",
          "impersonation",
          "inappropriate",
          "other"
        ]
      },
      "ReportStatus": {
        "type": "string",
        "enum": [
          "open",
          "actioned",
          "dismissed"
        ]
      },
      "ReviewReportDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reviewId": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReportReason"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReportStatus"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "reviewId",
          "reason",
          "status",
          "createdAt"
        ]
      },
      "ReportReviewDto": {
        "type": "object",
        "properties": {
          "reason": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReportReason"
              }
            ]
          },
          "note": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          }
        },
        "required": [
          "reason"
        ]
      },
      "ReviewReportEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReviewReportDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data"
        ]
      }
    }
  }
}